+++ Telegram/SourceFiles/inline_bots/inline_results_inner.cpp
@@ -16,7 +16,7 @@ https://github.com/telegramdesktop/tdesktop/blob/master/LEGAL
 #include "inline_bots/inline_bot_result.h"
 #include "inline_bots/inline_bot_layout_item.h"
 #include "lang/lang_keys.h"
-#include "layout/layout_utils.h"
+#include "layout/layout_position.h"
 #include "mainwindow.h"
 #include "facades.h"
 #include "main/main_session.h"
